Refactor Events
